Role CRUD alpha version

+++ app/Http/Controllers/Api/V1/Dashboard/RoleController.php
@@ -3,7 +3,11 @@
 namespace App\Http\Controllers\Api\V1\Dashboard;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Dashboard\StoreRoleRequest;
+use App\Http\Resources\Dashboard\RoleResource;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Role;
 
 class RoleController extends Controller
@@ -11,41 +15,43 @@ class RoleController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return RoleResource
      */
     public function index(Request $request)
     {
+//        abort_if(Gate::denies('role_access'), Response::HTTP_FORBIDDEN,'403 Forbidden');
         if (isset($request->limit)){
-            $role = Role::paginate($request->limit);
+            $role = Role::with(['permissions'])->paginate($request->limit);
         }
         else{
-            $role = Role::all();
+            $role = Role::with(['permissions'])->get();
         }
-        return $role;
+        return new RoleResource($role);
     }
 
     /**
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse|Response|object
      */
-    public function store(Request $request)
+    public function store(StoreRoleRequest $request)
     {
         $role = Role::create($request->validated());
-        return $role;
+        $role->permissions()->sync($request->input('permissions.*.id',[]));
+        return (new RoleResource($role))->response()->setStatusCode(Response::HTTP_CREATED);
     }
 
     /**
      * Display the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return RoleResource
      */
-    public function show($id)
+    public function show(Role $role)
     {
-        $role = Role::with('model_has_role')->findOrFail($id);
-        return $role;
+        abort_if(Gate::denies('role_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        return new RoleResource($role->load(['permissions']));
     }
 
     /**
@@ -53,25 +59,25 @@ class RoleController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse|Response|object
      */
     public function update(Request $request, Role $role)
     {
         $role->update($request->validated());
-        return $role;
+        $role->permissions()->sync($request->input('permissions.*.id',[]));
+        return (new RoleResource($role))->response()->setStatusCode(Response::HTTP_ACCEPTED);
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\JsonResponse
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\Routing\ResponseFactory|\Illuminate\Http\JsonResponse|Response
      */
     public function destroy(Role $role)
     {
+        abort_if(Gate::denies('role_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $role->delete();
-        return response()->json([
-            'message'=>'Successfully deleted!'
-        ]);
+        return response(null, Response::HTTP_NO_CONTENT);
     }
 }
